LRQA-42795 Create GitUtil and move ExecutionResult class into it
